src/py-client/requirements.txt
pynsq==0.8.1
